+++ packages/orq-rc/src/models/components/evaluatorresponsellm.ts
@@ -399,8 +399,8 @@ export const EvaluatorResponseLlm$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-01T12:26:04.887Z"),
-  updated: z.string().default("2026-07-01T12:26:04.887Z"),
+  created: z.string().default("2026-07-01T12:58:10.623Z"),
+  updated: z.string().default("2026-07-01T12:58:10.623Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.nullable(
     z.union([